requirements.txt
python-dotenv==1.2.1
ollama==0.6.1
chromadb==1.5.0
requests==2.32.5
colorama==0.4.6
duckduckgo-search==8.1.1
pytest==8.3.4
pytest-asyncio==1.3.0
websockets==16.0
fastapi==0.129.0
uvicorn==0.41.0
psutil==7.2.2
httpx==0.28.1
beautifulsoup4==4.13.3
pytest-playwright==0.7.0
